refactor: data 파일은 data 폴더 path 를 가지도록 함
