+++ front/components/Footer.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import Intersection from './recycle/Intersection';
-import { sidebarList } from './sidebar/ListName';
+import { sidebarList } from './sidebar/data/ListName';
 
 import { FaGithub, FaTwitter } from 'react-icons/fa';
