+++ front/components/details/data/ElementData.ts
