+++ front/components/recycle/submitSuccess/data/data.ts
