+++ front/components/sidebar/MobileList.tsx
@@ -7,7 +7,7 @@ import { logoutRequestAction } from '../../reducers/user';
 
 import { MdOutlineKeyboardArrowDown } from 'react-icons/md';
 
-import { sidebarList, dropList } from './ListName';
+import { sidebarList, dropList } from './data/ListName';
 
 const MobileList = () => {
   const [clickDrop, setClickDrop] = useState<boolean>(false);
